Added float support.

+++ Constant.cs
@@ -27,14 +27,14 @@ namespace LibNoise.Modifiers
     public class Constant
         : IModule
     {
-        public double Value { get; set; }
+        public float Value { get; set; }
 
-        public Constant(double value)
+        public Constant(float value)
         {
             Value = value;
         }
 
-        public double GetValue(double x, double y, double z)
+        public float GetValue(float x, float y, float z)
         {
             return Value;
         }
